+++ omisego-admin/src/main/java/co/omisego/omisego/model/params/AccountListParams.kt
@@ -7,8 +7,8 @@ package co.omisego.omisego.model.params
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
 
-import co.omisego.omisego.model.pagination.Filter
-import co.omisego.omisego.model.pagination.FilterableParams
+import co.omisego.omisego.model.filterable.Filter
+import co.omisego.omisego.model.filterable.FilterableParams
 import co.omisego.omisego.model.pagination.Paginable
 import co.omisego.omisego.model.pagination.PaginableParams
 import co.omisego.omisego.model.pagination.SortDirection
@@ -51,9 +51,6 @@ data class AccountListParams internal constructor(
 
     /**
      * A term to search for in all of the searchable fields.
-     * See more at [Paginable.Account.SearchableFields]
-     *
-     * Note: Conflict with searchTerms, only use one of them.
      */
     override val searchTerm: String? = null,
