+++ omisego-core/src/main/java/co/omisego/omisego/model/filterable/FilterableParams.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego.model.pagination
+package co.omisego.omisego.model.filterable
 
 /*
  * OmiseGO
